replace argument types with uint64_t

+++ src/algorithms/channel/libs/channel_msg_receiver_cc.cc
@@ -33,6 +33,7 @@
 #include <gnuradio/gr_complex.h>
 #include <gnuradio/io_signature.h>
 #include <glog/logging.h>
+#include <cinttypes>
 
 using google::LogMessage;
 
@@ -70,7 +71,7 @@ void channel_msg_receiver_cc::msg_handler_events(pmt::pmt_t msg)
                 d_channel_fsm->Event_failed_tracking_standby();
                 break;
             default:
-                d_channel_fsm->Event_set_reckernel_start(message);
+                d_channel_fsm->Event_set_reckernel_start(static_cast<uint64_t> message);
                 //LOG(WARNING) << "Default case, invalid message.";
                 break;
             }
